Fixed 10 more warnings, switched SMJobBlessTest to ARC rather than GC.
git-svn-id: https://svn.macports.org/repository/macports/branches/gsoc15-pallet@138382 d073be05-634f-4543-b044-5fe20cf6d1d6

+++ Pallet/PortsTableController.m
@@ -71,7 +71,7 @@
 #pragma mark Private Methods
 
 - (void)changePredicateEditorSize:(NSInteger) newRowCount {
-    NSLog(@"ROWS: %d", newRowCount);
+    NSLog(@"ROWS: %ld", (long)newRowCount);
     if (newRowCount == rowCount)
         return;
